working interpreter with Hello World! in the input.txt file

+++ stack.h
@@ -7,8 +7,8 @@ typedef void * Stack;
 Stack create_stack();
 void destroy_stack(Stack hStack);
 
-void stack_push(Stack *hStack, int data);
-void stack_pop(Stack *hStack, int *data);
+void stack_push(Stack *hStack, unsigned long long data);
+void stack_pop(Stack *hStack, unsigned long long *data);
 
 int stack_top(Stack *hStack);
 int is_stack_empty(Stack *hStack);
